Problem: logflare:sync doesn't return a response from Logflare
Solution: await for the status and the body of the response to return it

This changed a number of things. Firstly, original `sync` was not actually
waiting for the request to go through. Instead, it was waiting until the
request was made.

Now, we needed to handle info messages from Gun indicating the response status
and the body of the response to prepare a response. We used reversed order
batches before because it meant we don't need to reverse the batch once we assemble
it, however, this didn't quite work out for Gun messages as we needed to receive them
in the correct order. Thus, we no longer use `reversed_batch` in the batch server
and simply reverse the batch once it is processed.

This change also introduces a new dependency, jsone. It is used to parse JSON responses
for Logflare. This is to change in the future, when Logflare will respond with BERT
for BERT requests. At that point, this dependency can be dropped.

Fixes #1